Bitmask Username Update/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tweepy==4.14.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
requests-oauthlib==1.3.1
itsdangerous==2.1.2
pytest==7.4.3
pytest-asyncio==0.21.1

